Moved more global stuff into symbols.c for good. Lengths in symbols.h interface now size_t. Got rid of a long-standing compiler warning but had to add some dynamic checking with assertions.
git-svn-id: svn://svn.code.sf.net/p/dasm-dillon/code/trunk@253 77a3a24b-2c49-0410-90f5-8a0d088e77bf

+++ src/symbols.h
@@ -32,8 +32,8 @@
 
 void  setspecial(int value, int flags);
 SYMBOL *allocsymbol(void);
-SYMBOL *findsymbol(const char *str, int len);
-SYMBOL *CreateSymbol(const char *str, int len);
+SYMBOL *findsymbol(const char *str, size_t len);
+SYMBOL *CreateSymbol(const char *str, size_t len);
 void FreeSymbolList(SYMBOL *sym);
 void programlabel(void);
 void debug_symbol_hash_collisions(void);
